pyproject.toml
[tool.poetry]
name = "ocr-toolkit"
version = "1.0.0"
description = "OCR Toolkit API for text extraction and language detection"
authors = ["OCR Toolkit <support@ocr-toolkit.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
isort = "^5.12.0"
pytest = "^7.4.0"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
generate-docs = "python generate_openapi.py"
